Abstract the Reader factory and implement reader for textual files
