+++ source/ObjectViewer/ObjectManager.cs
@@ -1671,7 +1671,7 @@ namespace OpenBve
 		            Result = WavefrontObjParser.ReadObject(FileName, Encoding, LoadMode, ForceTextureRepeatX, ForceTextureRepeatY);
 		            break;
 				case ".s":
-					Result = BinaryShapeParser.ReadObject(FileName);
+					Result = MsTsShapeParser.ReadObject(FileName);
 					break;
 					default:
                     Interface.AddMessage(Interface.MessageType.Error, false, "The file extension is not supported: " + FileName);
